Facial-Expression-Recognition-main/requirements.txt
h5py
opencv-python
keras
matplotlib
numpy
opencv-contrib-python
pandas
Pillow
protobuf
streamlit
tensorboard
tensorflow
zipp
